+++ TextField/TextField.jsx
@@ -74,6 +74,9 @@ class TextFieldInput extends MaterialComponent {
   componentWillUnmount() {
     this.MDComponent && this.MDComponent.destroy && this.MDComponent.destroy();
   }
+  value() {
+    return this.MDComponent?this.MDComponent.value:null;
+  }
   materialDom(allprops) {
     let {className, ...props} = allprops;
     className = className || '';